added listeners and tested existing listners
main thing i did here was work on the listener after the activity is created. I treated this area like a main, which i dont think is correct, but i just wanted to get some functionality. commenting is low, but i can explain in class. everthing here is fairly straight forward. The other big change that was made in this was that i changed activities from a string arraylist to a node arraylist. the reason i did this was so i could get the name all the nodes